fix(tests): point edge path guards at src/ layout after edge restructure
